Allow PHPUnit_Framework_TestSuite subclasses.
PHPUnit expects PHPUnit_Framework_TestSuite subclasses to be added to an
existing test suite as real objects rather than as ReflectionClass
wrappers. This small change BatchTest checks the type of each test
before passing it on to PHPUnit and skips the reflection wrapper if the
class is of the correct type.