Fix unit tests.
They only used to work because the relative symlinks were resolved using
the current working directory as the point of origin, while it should have
been the symlink's directory.

Commenting the "clean" target in test/etc/tasks/system/AvailableTaskTest.xml
shows that the symlinks created there were actually dangling, but the
resolution algorithm failed to produce the correct path and therefore did not
notice that.