Add two new PropertyTask tests where circular definition detection fails.
Below, A->B means resolves to B in some way, like A=xx${B}yy

Currently a chain like A->B->A will be detected in the
resolveAllProperties() function since we check each new property in the
chain against the property were trying to resolve.

This fails if we start to resolve A but the circular chain doesn't
involve A:
A->B->C->B->C->B....

These two tests reflect that.