Fix tests.
Honestly, AbstractFileSetTest is a little mess. Seems that previously
just calling createInclude() would suffice to end up with a not-empty patternset.

With the current changes, where the include can contain a comma-separated list of pattern names,
not setting the name will result in not having set a pattern at all.