ChmodTaskTest: remove creation of extra "chmodtest" file so that tearDown() can delete the directory
The "chmodtest" file had no purpose for the test. It was probably just copy-pasted from different test case. When the file was present inside the "A" directory it was not possible clean it up because the directory has only "r" permission after finishing the test (therefore files inside the directory cannot be removed).