MkdirTask behaves the same as "mkdir" Linux command and respects POSIX ACL (#591)
* MkdirTask behaves the same as "mkdir" Linux command and respects ACL

Previously, createDirectory() method had default value of the $mode argument 0755. This did not work well with POSIX ACL, because if the $mode argument was not provided the newly created directory did not have the exact ACL permissions inherited from default ACL of the parent directory (as when calling "mkdir" console command without "--mode" option), but this permissions were narrowed by the 0755 mask.

For more infomation about how mode argument of Unix mkdir() function behaves when using POSIX ACL see "OBJECT CREATION AND DEFAULT ACLs" here: https://linux.die.net/man/5/acl

* PhingFile: remove unused $mode argument from createNewFile() method

* Added MkdirTaskTest

* PhingFile::mkdirs() sets the same permissions as "mkdir -p" Linux command

Parent directories should be created with the default permissions and not the ones passed in "mode" argument.

* ChmodTaskTest: remove creation of extra "chmodtest" file so that tearDown() can delete the directory

The "chmodtest" file had no purpose for the test. It was probably just copy-pasted from different test case. When the file was present inside the "A" directory it was not possible clean it up because the directory has only "r" permission after finishing the test (therefore files inside the directory cannot be removed).

* travis.yml: Enable ACL support in the Travis CI VM

* remove extra require_once + add license header

* document the BC breaks in UPGRADING.md

* Update MkdirTaskTest to skip scenarios that use non-zero umask on Windows

Windows does not support umask. On Windows, umask behaves as if it
was hardcoded to 0000.
Also, add some scenarios that do not depend on umask that will be run
both on Linux and Windows.

* Update MkdirTaskTest to use test data that work both on Linux and Windows

* properly clean-up even if test did not create any directory

* Last try before I rage quit and go to bed

* Update MkdirTaskTest.php

* Update MkdirTaskTest.php

Co-authored-by: Siad Ardroumli <siad.ardroumli@gmail.com>
Co-authored-by: Siad Ardroumli <siad.ardroumli@googlemail.com>