Call vsprintf correctly in ChmodTask (#1627)
`vsprint` expects its second argument to be an array. Old PHP versions silently also accepted a single non-array parameter that was "upgraded" to [$param].
This becomes an issue on PHP8, because it type-checks the parameter and issues an error accordingly.

This issue is already fixed in `main` and only exists in `oldstable`.