Fix Grammar: composerTask attributes are each optional
According to docu https://www.phing.info/guide/hlhtml/#ComposerTask composer and php attributes are optional, with default behaviour if they are omitted. 
It works to use one of them.
The old version of the grammar did not allow only one of them - only none or both. 
With this fix both, none or one of them are allowed and correctly validated